tests/wpt/tests/quirks/table-cell-width-calculation-applies-to.html

<title>The table cell width calculation quirk should not apply to inline blocks</title>
<link rel="author" title="Koji Ishii" href="mailto:kojii@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<style>
inline-block {
  display: inline-block;
}
</style>
<body>
  <table style="width:80px" data-expected-width=80>
    <tr>
      <td><inline-block style="width:50px">50</inline-block><inline-block style="width:50px">50</inline-block></td>
    </tr>
  </table>
<script>checkLayout('table');</script>
</body>
